+++ quasar-utils/src/main/groovy/net/corda/plugins/quasar/QuasarPlugin.groovy
@@ -1,14 +1,17 @@
 package net.corda.plugins.quasar
 
 import groovy.transform.PackageScope
+import org.gradle.api.Action
 import org.gradle.api.GradleException
 import org.gradle.api.InvalidUserDataException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.ConfigurationContainer
+import org.gradle.api.artifacts.DependencySet
+import org.gradle.api.artifacts.dsl.DependencyHandler
 import org.gradle.api.model.ObjectFactory
-import org.gradle.api.plugins.JavaPlugin
+import org.gradle.api.provider.Property
 import org.gradle.util.GradleVersion
 
 import static org.gradle.api.plugins.JavaPlugin.COMPILE_ONLY_CONFIGURATION_NAME
@@ -29,7 +32,7 @@ class QuasarPlugin implements Plugin<Project> {
     private static final String CORDA_PROVIDED_CONFIGURATION_NAME = "cordaProvided"
     private static final String CORDA_RUNTIME_ONLY_CONFIGURATION_NAME = "cordaRuntimeOnly"
     @PackageScope static final String defaultGroup = "co.paralleluniverse"
-    @PackageScope static final String defaultVersion = "0.8.2_r3"
+    @PackageScope static final String defaultVersion = "0.8.4_r3"
 
     private final ObjectFactory objects
 
@@ -44,23 +47,20 @@ class QuasarPlugin implements Plugin<Project> {
             throw new GradleException("The Quasar-Utils plugin requires Gradle $MINIMUM_GRADLE_VERSION or newer.")
         }
 
-        // Apply the Java plugin on the assumption that we're building a JAR.
-        // This will also create the "compileOnly" and "runtimeOnly" configurations.
-        project.pluginManager.apply(JavaPlugin)
-
         def rootProject = project.rootProject
-        def quasarGroup = rootProject.hasProperty('quasar_group') ? rootProject.property('quasar_group') : defaultGroup
-        def quasarVersion = rootProject.hasProperty('quasar_version') ? rootProject.property('quasar_version') : defaultVersion
-        def quasarPackageExclusions = rootProject.hasProperty("quasar_exclusions") ? rootProject.property('quasar_exclusions') : Collections.emptyList()
+        def quasarGroup = rootProject.findProperty('quasar_group')?.toString() ?: defaultGroup
+        def quasarVersion = rootProject.findProperty('quasar_version')?.toString() ?: defaultVersion
+        def quasarSuspendable = rootProject.findProperty('quasar_suspendable_annotation')?.toString()?.trim()
+        def quasarPackageExclusions = rootProject.findProperty('quasar_exclusions') ?: Collections.emptyList()
         if (!(quasarPackageExclusions instanceof Iterable<?>)) {
             throw new InvalidUserDataException("quasar_exclusions property must be an Iterable<String>")
         }
-        def quasarClassLoaderExclusions = rootProject.hasProperty("quasar_classloader_exclusions") ? rootProject.property('quasar_classloader_exclusions') : Collections.emptyList()
+        def quasarClassLoaderExclusions = rootProject.findProperty('quasar_classloader_exclusions') ?: Collections.emptyList()
         if (!(quasarClassLoaderExclusions instanceof Iterable<?>)) {
             throw new InvalidUserDataException("quasar_classloader_exclusions property must be an Iterable<String>")
         }
         def quasarExtension = project.extensions.create(QUASAR, QuasarExtension, objects,
-                quasarGroup, quasarVersion, quasarPackageExclusions, quasarClassLoaderExclusions)
+            quasarGroup, quasarVersion, quasarSuspendable, quasarPackageExclusions, quasarClassLoaderExclusions)
 
         addQuasarDependencies(project, quasarExtension)
         configureQuasarTasks(project, quasarExtension)
@@ -83,16 +83,15 @@ class QuasarPlugin implements Plugin<Project> {
             dependencies.add(quasarAgentDependency)
         }
 
-        // Add Quasar bundle to the compile classpath WITHOUT any of its transitive dependencies.
-        createCompileOnlyConfiguration(CORDA_PROVIDED_CONFIGURATION_NAME, project.configurations).extendsFrom(quasar)
+        // If we're building a JAR then also add the Quasar bundle to the appropriate configurations.
+        project.pluginManager.withPlugin('java') {
+            // Add Quasar bundle to the compile classpath WITHOUT any of its transitive dependencies.
+            def cordaProvided = createCompileOnlyConfiguration(CORDA_PROVIDED_CONFIGURATION_NAME, project.configurations)
+            cordaProvided.withDependencies(new QuasarAction(project.dependencies, extension, false))
 
-        // Instrumented code needs both the Quasar bundle and its transitive dependencies at runtime.
-        def cordaRuntimeOnly = createRuntimeOnlyConfiguration(CORDA_RUNTIME_ONLY_CONFIGURATION_NAME, project.configurations)
-        cordaRuntimeOnly.withDependencies { dependencies ->
-            def quasarDependency = project.dependencies.create(extension.dependency.get()) {
-                it.transitive = true
-            }
-            dependencies.add(quasarDependency)
+            // Instrumented code needs both the Quasar bundle and its transitive dependencies at runtime.
+            def cordaRuntimeOnly = createRuntimeOnlyConfiguration(CORDA_RUNTIME_ONLY_CONFIGURATION_NAME, project.configurations)
+            cordaRuntimeOnly.withDependencies(new QuasarAction(project.dependencies, extension, true))
         }
     }
 
@@ -131,5 +130,31 @@ class QuasarPlugin implements Plugin<Project> {
             }
         }
         return configuration
+    }
+
+    private static class QuasarAction implements Action<DependencySet> {
+        private final DependencyHandler handler
+        private final QuasarExtension extension
+        private final boolean isTransitive
+
+        QuasarAction(DependencyHandler handler, QuasarExtension extension, boolean isTransitive) {
+            this.handler = handler
+            this.extension = extension
+            this.isTransitive = isTransitive
+        }
+
+        @Override
+        void execute(DependencySet dependencies) {
+            if (isEmpty(extension.suspendableAnnotation)) {
+                def quasarDependency = handler.create(extension.dependency.get()) { dep ->
+                    dep.transitive = isTransitive
+                }
+                dependencies.add(quasarDependency)
+            }
+        }
+
+        private static boolean isEmpty(Property<String> property) {
+            !property.isPresent() || property.get().isBlank()
+        }
     }
 }